Task12 Upd (data loading once)
