+++ src/pages/Root.js
@@ -1,6 +1,5 @@
-import axios from "axios";
 import React from "react";
-import { json, Outlet, useLoaderData } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import MainNavigation from "../components/Layout/MainNavigation";
 
 const RootLayout = () => {
